+++ src/main/java/com/erebonk/elasticsearch/service/parser/impl/ParserServiceImpl.java
@@ -15,7 +15,6 @@ import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.io.InputStream;
 import java.math.BigDecimal;
-import java.net.URL;
 import java.util.Iterator;
 import java.util.stream.Stream;
