--- src/main/resources/application-production.yml
@@ -1,4 +1,0 @@
-spring:
-  elasticsearch:
-    rest:
-      uris: http://localhost:9200
